removed many int casts and changed the value to float (for improved precision)

+++ Baselet/src/com/baselet/control/Utils.java
@@ -83,7 +83,7 @@ public abstract class Utils {
 		return returnVector;
 	}
 
-	public static List<String> splitString(String text, double width) {
+	public static List<String> splitString(String text, float width) {
 		DiagramHandler handler = Main.getInstance().getDiagramHandler();
 		StringBuilder stringBuilder = new StringBuilder(text);
 		int lastEmptyChar = -1; // is -1 if there was no ' ' in this line
@@ -128,7 +128,7 @@ public abstract class Utils {
 		return ret;
 	}
 
-	public static Stroke getStroke(LineType lineType, int lineThickness) {
+	public static Stroke getStroke(LineType lineType, float lineThickness) {
 		// If the lineThickness is not supported, the default type is used
 		if (lineThickness < 0) lineThickness = Constants.DEFAULT_LINE_THICKNESS;
